refactor: move HOCs to `helpers` folder
