+++ ts/features/itwallet/wallet/screens/ItwOfflineWalletScreen.tsx
@@ -1,9 +1,9 @@
 import { HeaderFirstLevel } from "@pagopa/io-app-design-system";
 import { IOScrollView } from "../../../../components/ui/IOScrollView";
 import I18n from "../../../../i18n";
-import { ItwWalletCardsContainer } from "../components/ItwWalletCardsContainer.tsx";
-import { withOfflineAlert } from "../utils/withOfflineAlert";
-import { useOnFirstRender } from "../../../../utils/hooks/useOnFirstRender.ts";
+import { ItwWalletCardsContainer } from "../components/ItwWalletCardsContainer";
+import { withOfflineAlert } from "../../common/helpers/withOfflineAlert";
+import { useOnFirstRender } from "../../../../utils/hooks/useOnFirstRender";
 import { trackItwOfflineWallet } from "../../analytics";
 
 const ItwOfflineWalletScreen = () => {
